ocrvs-5514 death & marriage form refactor (#609)
* refactor common fields

* add introduction for death

* deceased page set

* refactor death event and location

* refgactor informant page of death form

* refactor and clean whole form

* hidden registration section add

* refactor marriage informant page

* deprecate contact point for marriage

* refactor groom's details age

* refactor bride page

* refactor marriage details page

* refactor witness page

* refactor marriage doc uploader page

* informant section add for marriage form

* update import and clean up ts error

* form message ts ignore and import update

* Reorganize the util functions

---------
