+++ src/features/languages/content/client/client.json
@@ -1209,6 +1209,7 @@
         "form.section.deceased.mother.name": "What is the deceased's mother name?",
         "form.section.deceased.mother.title": "Mother's details",
         "form.section.deceased.name": "Deceased",
+        "form.section.deceased.relationship": "Relationship to deceased",
         "form.section.deceased.noSpouse": "No / Unknown",
         "form.section.deceased.spouse.name": "Does the deceased have a spouse?",
         "form.section.deceased.spouse.title": "Spouse's details",
@@ -1219,6 +1220,10 @@
         "form.section.information.birth.bullet2": "As the legal Informant it is important that all the information provided by you is accurate.",
         "form.section.information.birth.bullet3": "Once the declaration is processed you will receive you will receive an SMS to tell you when to visit the office to collect the certificate - Take your ID with you.",
         "form.section.information.birth.bullet4": "Make sure you collect the certificate. A birth certificate is critical for this child, especially to make their life easy later on. It will help to access health services, school examinations and government benefits.",
+        "form.section.information.death.bullet1": "I am going to help you make a declaration of death.",
+        "form.section.information.death.bullet2": "As the legal Informant it is important that all the information provided by you is accurate.",
+        "form.section.information.death.bullet3": "Once the declaration is processed you will receive you will receive an SMS to tell you when to visit the office to collect the certificate - Take your ID with you.",
+        "form.section.information.death.bullet4": "Make sure you collect the certificate. A death certificate is critical to support with inheritance claims and to resolve the affairs of the deceased e.g. closing bank accounts and setting loans.",
         "form.section.declaration.title": "Declaration details",
         "form.section.documents.birth.requirements": "The following documents are required",
         "form.section.documents.list.attestedBirthRecord": "Attested copy of hospital document or birth record, or",
@@ -1363,6 +1368,7 @@
         "form.section.marriageEvent.name": "Marriage event details",
         "form.section.marriageEvent.title": "Marriage details",
         "form.field.label.marriedLastName": "Married last name (if different)",
+        "form.field.label.lastNameAtBirth": "Last name at birth (if different from above)",
         "form.field.label.docTypeMarriageNotice": "Notice of marriage",
         "form.field.label.proofOfGroomsID": "Proof of groom's identity",
         "form.field.label.proofOfBridesID": "Proof of bride's identity",
@@ -1672,10 +1678,7 @@
         "register.eventInfo.birth.title": "Introduce the birth registration process to the informant",
         "register.eventInfo.death.title": "Introduce the death registration process to the informant",
         "register.eventInfo.marriage.title": "Introduce the marriage registration process to the informant",
-        
-        
-        
-        
+
         "register.form.informationMissing": "Information missing",
         "register.form.missingFieldsDescription": "The following information will be submitted for validation. Please\n    make sure all required details have been filled in correctly. There\n    are {numberOfErrors} missing mandatory fields in your form:",
         "register.form.modal.areYouReadyToSubmit": "Are you ready to submit?",
@@ -3063,7 +3066,7 @@
         "form.field.label.fetchRegistrationModalErrorText": "Aucun enregistrement trouvé pour le BRN fourni",
         "form.field.label.fetchRegistrationModalInfo": "Numéro d'enregistrement de naissance",
         "form.field.label.fileUploadError": "{type} supporté uniquement",
-        "form.field.label.fileSizeError" : "La taille du fichier doit être inférieure à 2 Mo",
+        "form.field.label.fileSizeError": "La taille du fichier doit être inférieure à 2 Mo",
         "form.field.label.firstName": "Prénom",
         "form.field.label.firstNames": "Prénom(s)",
         "form.field.label.firstNameEN": "Prénom anglais",
@@ -3283,6 +3286,7 @@
         "form.section.deceased.mother.name": "Quel est le nom de la mère de la personne décédée ?",
         "form.section.deceased.mother.title": "Détails de la mère",
         "form.section.deceased.name": "Défunt",
+        "form.section.deceased.relationship": "Relation avec le défunt",
         "form.section.deceased.noSpouse": "Non / Inconnu",
         "form.section.deceased.spouse.name": "Le défunt a-t-il un conjoint ?",
         "form.section.deceased.spouse.title": "Coordonnées du conjoint",
@@ -3293,6 +3297,10 @@
         "form.section.information.birth.bullet2": "En tant qu'informateur légal, il est important que toutes les informations que vous fournissez soient exactes.",
         "form.section.information.birth.bullet3": "Une fois la déclaration traitée, vous recevrez un SMS vous indiquant quand vous rendre au bureau pour retirer le certificat - Munissez-vous d'une pièce d'identité.",
         "form.section.information.birth.bullet4": "Veillez à récupérer le certificat. Un certificat de naissance est essentiel pour cet enfant, notamment pour lui faciliter la vie plus tard. Il l'aidera à accéder aux services de santé, aux examens scolaires et aux prestations de l'État.",
+        "form.section.information.death.bullet1": "Je vais vous aider à faire une déclaration de décès.",
+        "form.section.information.death.bullet2": "En tant qu'informateur légal, il est important que toutes les informations que vous fournissez soient exactes.",
+        "form.section.information.death.bullet3": "Une fois la déclaration traitée, vous recevrez un SMS vous indiquant quand vous rendre au bureau pour retirer le certificat - Munissez-vous d'une pièce d'identité.",
+        "form.section.information.death.bullet4": "Veillez à récupérer le certificat. Le certificat de décès est essentiel pour les demandes d'héritage et pour régler les affaires de la personne décédée, par exemple la fermeture des comptes bancaires et la mise en place des prêts.",
         "form.section.declaration.title": "Détails de la déclaration",
         "form.section.documents.birth.requirements": "Pour l'enregistrement de la naissance, les documents suivants sont requis :",
         "form.section.documents.list.attestedBirthRecord": "Copie certifiée conforme du document de l'hôpital ou de l'acte de naissance, ou",
@@ -3435,6 +3443,7 @@
         "form.section.marriageEvent.name": "Détails de l'événement de mariage",
         "form.section.marriageEvent.title": "Détails du mariage",
         "form.field.label.marriedLastName": "Marié Nom de famille (si différent)",
+        "form.field.label.lastNameAtBirth": "Nom de famille à la naissance (si différent de ci-dessus)",
         "form.field.label.docTypeMarriageNotice": "Avis de mariage",
         "form.field.label.proofOfGroomsID": "Preuve de l'identité du marié",
         "form.field.label.proofOfBridesID": "Preuve de l'identité de la mariée",
